chore: minor improvements
