+++ apps/api/src/controllers/admin/manage/AdminManageCitizensController.ts
@@ -7,15 +7,15 @@ import { prisma } from "lib/prisma";
 import { IsAuth } from "middlewares/IsAuth";
 import { CREATE_CITIZEN_SCHEMA } from "@snailycad/schemas";
 import { validateSchema } from "lib/validateSchema";
-import { generateString } from "utils/generateString";
+import { generateString } from "utils/generate-string";
 import { citizenInclude } from "controllers/citizen/CitizenController";
-import { validateImgurURL } from "utils/images/image";
+import { validateImageURL } from "lib/images/validate-image-url";
 import { Prisma, Rank } from "@prisma/client";
 import { UsePermissions, Permissions } from "middlewares/UsePermissions";
 import { isCuid } from "cuid";
 import type * as APITypes from "@snailycad/types/api";
 import { validateSocialSecurityNumber } from "lib/citizen/validateSSN";
-import generateBlurPlaceholder from "utils/images/generate-image-blur-data";
+import generateBlurPlaceholder from "lib/images/generate-image-blur-data";
 import { leoProperties, unitProperties } from "lib/leo/activeOfficer";
 import { AuditLogActionType, createAuditLogEntry } from "@snailycad/audit-logger/server";
 
@@ -140,7 +140,7 @@ export class AdminManageCitizensController {
       });
     }
 
-    const validatedImageURL = validateImgurURL(data.image);
+    const validatedImageURL = validateImageURL(data.image);
 
     const updatedCitizen = await prisma.citizen.update({
       where: { id },
