+++ apps/api/src/controllers/admin/manage/cad-settings/CADImagesController.ts
@@ -11,7 +11,7 @@ import { cad, Rank } from "@prisma/client";
 import { Permissions } from "@snailycad/permissions";
 import { UsePermissions } from "middlewares/UsePermissions";
 import { ExtendedBadRequest } from "src/exceptions/ExtendedBadRequest";
-import { getImageWebPPath } from "utils/images/image";
+import { getImageWebPPath } from "lib/images/validate-image-url";
 
 @Controller("/admin/manage/cad-settings/image")
 @ContentType("application/json")
