+++ apps/api/src/controllers/bleeter/BleeterController.ts
@@ -20,9 +20,9 @@ import { validateSchema } from "lib/validateSchema";
 import { ExtendedBadRequest } from "src/exceptions/ExtendedBadRequest";
 import type { User } from "@prisma/client";
 import type * as APITypes from "@snailycad/types/api";
-import { getImageWebPPath } from "utils/images/image";
+import { getImageWebPPath } from "lib/images/validate-image-url";
 import { Feature, IsFeatureEnabled } from "middlewares/is-enabled";
-import generateBlurPlaceholder from "utils/images/generate-image-blur-data";
+import generateBlurPlaceholder from "lib/images/generate-image-blur-data";
 
 @UseBeforeEach(IsAuth)
 @Controller("/bleeter")
