+++ apps/api/src/controllers/record/records-controller.ts
@@ -33,7 +33,7 @@ import { getFirstOfficerFromActiveOfficer, getInactivityFilter } from "lib/leo/u
 import type * as APITypes from "@snailycad/types/api";
 import { officerOrDeputyToUnit } from "lib/leo/officerOrDeputyToUnit";
 import { Socket } from "services/socket-service";
-import { assignUnitsToWarrant } from "lib/records/assignToWarrant";
+import { assignUnitsToWarrant } from "lib/records/assign-units-to-warrant";
 import type { cad } from "@snailycad/types";
 import { userProperties } from "lib/auth/getSessionUser";
 import { upsertRecord } from "lib/records/upsert-record";
