+++ apps/api/src/utils/check-for-updates.ts
