+++ apps/api/src/utils/generate-string.ts
